fix Complete method to hand empty field name
Previously the implementation of Complete used the zero
value of a string to detect and handle the first time
though a loop iteration.  If however, we decide to support
the empty field name, this change is required to correctly
differentiate between first iteration of the loop and a
field with empty name.

relates to #1593