introduce new hierarchy token filter (#1570)
* introduce new hierarchy token filter

This token filter takes a value representing hierarchical
informtion separated by a delimiter and produces tokens
for each level of the hierarchy.

For example, consider a field with the value "a/b/c".
This token filter is useful when the interpretation of
this is "a" is the most general categorization, "a/b" is more
specific, and finally "a/b/c" is the most specific.  Often
times one may wish to query for documents satisfying "a" or
"a/b".  This token filter generates tokens for these values
so that an exact term query can be used at search time.

Required configuration:

delimter - used for splitting (optional) and joining the parts

Option configuration:

max - the maximum number of parts to join

splitInput - if each input term should be split by the delimiter,
this defaults to true, allowing simpler usage by pairing it with
the single tokenizer

* fix copy/paste mistakes in the config