MB-58034: Enable datetime parser customization in the date range query object (#1860)
- Add a new query DateRangeStringQuery, which takes in the start and end
bounds as strings and then parses them using the DateTimeParser
set in the DateRangeStringQuery struct, if present and defaults to using the package level
QueryDateTimeParser. 
- ParseQuery, which analyzes a marshalled json to identify the query to
run, will now default to DateRangeStringQuery instead, which is preferable because it already has the dates in string format, and unmarshalling to DateRangeQuery will always trigger parsing of the date string with only QueryDateTimeParser, leaving no room for user configuration.