MB-58388: Use custom datetime parsers in date range facet queries (#1878)
Related to Date range facet queries:

- The "start" and "end" fields of the Facet Result struct are now
formatted with the layout that successfully parsed them in the Facet
Request struct. Earlier we always used RFC3339Nano by default which led
to disparity between the corresponding fields in the facet request and
result.
   Ex - Facet Request Date Range:

          { 
              start : "2014-09-10"
              end : "2015-08-11"
          }

Previously the Facet result had 

          { 
              start : "2014-09-10T00:00:00Z"
              end : "2015-08-11T00:00:00Z"
              Count: 100
          }

Now it has 

          { 
              start : "2014-09-10"
              end : "2015-08-11"
              Count: 100
          }
- Allow a custom date time parser to be added to the date range struct
of a facet request, that will override the default date time parser
(dateTimeOptional) while parsing the start and end date strings.
- Add error message when the date time parser fails to parse the date
string. Earlier if the date time parser failed to parse the start and
end strings of the date range in the facet request, no error was thrown
and the facet request would be for start and end being equal to 0 time,
so facet result was always nil.
- Fix marshalling error for dateRangeQuery struct. If the query had no
end / endString, which is the case when upper bound is unspecified, the
marshalled JSON always had the 0 time (time.Time{}) as the value for
end, now it is omitted from the marshalled JSON.