MB-60241: Add consistency checks for redistribution of KNN hits (#1958)
In a distributed indexing setup, the child indexes linked to an alias
are labeled as `[IP]:[port]/[indexName]`. During a two-phase KNN search,
if a node is taken out and re-added (by means of a rebalance operation)
after the first phase but before the start of the second phase, the
reference to the child index in the alias remains unchanged (with the
same IP address, port, and indexName). However, the actual indexes
present on that node will differ from those available during the first
phase. This discrepancy causes inaccurate results due to inconsistencies
in the indexId stack.

This method substitutes the indexId with the specific indexName,
ensuring redistribution even if the index arrangement undergoes changes.
Yet, if the index isn’t found within that alias, an error will be
triggered.

---------

Co-authored-by: Abhinav Dangeti <abhinav@couchbase.com>