Fix issue in parsing datetime field (#2030)
Addresses https://github.com/blevesearch/bleve/issues/2027

In bleve v2.3.10, [a bug
fix](https://github.com/blevesearch/bleve/pull/1868) was introduced
where the value of a datetime field in the Fields part of a search hit
was made to match the value actually present in the document.
- For example if a document A was indexed and had value of a datetime
field X to be "2001/08/20 03:00:10", and ifA was returned as part of the
search result, A's Fields will have a map which would look like
- Before Fix 
    - "X" : "2001-08-20T03:00:10Z"
- After fix 
    - "X" : "2001/08/20 03:00:10"

This would come to play in case of custom user defined date time parser,
and we did this by storing the layout with which the time string was
parsed with during indexing. With timestamp support being added at the
same time, they did not really have a valid layout to store and hence
stored nil. During queries, the layout extracted, if null, was always
assumed to be for the timestamp case _but_ the case where an existing
index from pre v2.3.10 (which would also return null layout) was not
handled and it became a catchall case leading to the datetime string
being always returned in a timestamp rather than the expected default
RFC3339.