MB-60945 - Add scorch event for pre merge check (#1971)
This PR adds an event to check whether a merge can proceed based on
criteria set by the caller(eg. cbft). If not, it's aborted and the
merger loop continues its next iteration.
Also includes minor refactor to the function signature of the event
callback registry to accommodate the new event type.