feat: remove sloppy math for sin, cos and asin
Remove the LUT implementation for `sin`, `cos` and `asin`. For `sin` and
`cos` they are not faster than the hardware provided `sin` and `cos`
instructions on AMD64 and ARM64. For `asin` LUT does seem to be a bit
faster but hardware and LUT implementation are both in the same order of
magnitude for performance and `asin` LUT unconditionally takes 1.5MiB of
memory which seems a bit excessive.