Introduce a new stat which indicates a bleve index's convergence (#2152)
`index_bgthreads_active` indicates whether the background routines that
maintain the index are busy doing some work. This means that the index
hasn't converged to a steady state and there could be potential file
segment merges or in-memory segment flushes still remaining.
This stat is beneficial for the application layer to get better insight
as to whether the index is doing some background work (which can have
implications on the system's resource utilisation) or not.