MB-62182: Making the merge planner aware of segment file size (#2134)
This PR introduces a new option of FloorSegmentFileSize into the merge planner's algorithm which is used in the budget calculation. Currently the budget calculation is dependent on the number of documents however this doesn't translate well in terms of resource utilization when the per doc size - which directly translates to the segment file size, varies significantly across use-cases and making the merger aware of the file size helps in addressing this issue.