MB-60971: update the bolt with the correct file segments (#2178)
- This is applicable to concurrent persister algorithm (introduced in
v2.5.0) - which is not _ON_ by default.
- In cases where the persister creates new file segments, the bolt file
is supposed to be updated with the complete set of file segments.
- However currently, while loading the index from the bolt source, there
can be data loss issues because a set of segments associated with the
root snapshot weren't tracked in the bolt file and hence wouldn't be
loaded from the disk.
- Note that the live index doesn't have issues because the root would
track the live persisted segments and point to the corresponding files.
- Also added some better code comments