MB-66163: fix checkpointing mechanism under sparse mutation scenarios  (#2186)
- the checkpointing mechanism considers snapshots which fall within the
`numSnapshotsToKeep * rollbackSamplingInterval` duration window. The
snapshots outside this window are purged.
- Under sparse mutation scenario for eg, when the workload pattern in
such that it results in seldomly persisting snapshots that too outside
this duration window - we could end up purging ALL those snapshots which
makes the system fallback to preserving the latest snapshots
- For a better partial rollback behaviour, we retain
`rollbackRetentionFactor` worth of snapshots out of this duration window
such that the cost of rebuilding the index from scratch is avoided.
- This PR fixes the decision logic involved in the scenario pointed and
tracks the previous checkpoints while opening back the index as well.