MB-65170: Add filter support to BooleanQuery (#2220)
- Add a Filter query for Boolean queries, which filters the document set
returned by the
Boolean query itself.
- The filter query does not affect scores, and any document returned by
the Boolean query
must also satisfy the filter query.
- The key difference between using a Filter query and placing an
equivalent query in the
Must clause is that queries in the Must clause contribute to the score.
The Filter
query is intended purely for filtering purposes without modifying the
document scores
set by the base Boolean query.
- Requires a Bug Fix (https://github.com/blevesearch/bleve/pull/2146)
to allow for the usage of the Advanced API in case the FilterQuery is an
optimized
composite searcher.
- Resolves https://github.com/blevesearch/bleve/issues/2037