MB-60401: Support RRF for hybrid search (#2218)
This PR introduces hybrid search into bleve. This comes with 2 major
additions: the `fusion` package and the `rescorer` interface.
- The `fusion` package is purely algorithmic to apply hybrid search
strategies (for this time, only reciprocal rank fusion) to a
`DocumentMatchCollection`. Additional algorithms can be added to the
package.
- The `rescorer` interface provides the compatibility layer to enable
hybrid search with bleve. This is done with the concrete type
`fusionRescorer` which applies the algorithms in `fusion` to the search
pipeline.